+++ srcs/expansion/expand_wildcard.c
@@ -6,7 +6,7 @@
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 10:28:30 by brappo            #+#    #+#             */
-/*   Updated: 2024/04/18 16:56:12 by brappo           ###   ########.fr       */
+/*   Updated: 2024/04/18 17:01:57 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
